01-cube done

+++ examples/00-simple/App.cpp
@@ -57,7 +57,7 @@ App::App()
     geometry.setTexCoords(texCoords);
 }
 
-void App::update(double deltaTime)
+void App::update(float deltaTime)
 {
 }
